Working basis
Chiara's code has been cut down to the basics and rewritten for AOD. Note that ESD and AOD seem to have different formats so ESD can't be used as an input in an AOD analysis. It's not that the input handler sifts through the ESD and pulls out AOD. Just for future reference.